requirements.txt
flask
python-dotenv
requests
supabase
authlib
flask-login